+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Directory: ForcePermanent option
+
 ## v2.20.0 - 2023-10-25
 
 ### Changed
+
 - datastore: compute number of samples in a trial at write time instead of at query time (includes an automatic migration).
 - datastore: pass samples to persistence layer asynchronously